+++ api/pin_reset.go
@@ -32,10 +32,12 @@ const (
 // @Param userid path string true "user id"
 // @Success 200 {string} string "OK"
 // @Failure 400 {object} status.Status "userId was not provided"
+// @Failure 401 {object} status.Status "only authorized for token bearers"
 // @Failure 403 {object} status.Status "only authorized for patients, not clinicians nor server token"
 // @Failure 422 {object} status.Status "Error when sending the email (probably caused by the mailing service)"
 // @Failure 500 {object} status.Status "Internal error while processing the request, detailed error returned in the body"
 // @Router /send/pin-reset/{userid} [post]
+// @security TidepoolAuth
 func (a *Api) SendPinReset(res http.ResponseWriter, req *http.Request, vars map[string]string) {
 
 	// by default, language is EN. It will be overriden if prefered language is found later
@@ -44,10 +46,9 @@ func (a *Api) SendPinReset(res http.ResponseWriter, req *http.Request, vars map[
 	var usrDetails *shoreline.UserData
 	var err error
 
-	var token = req.Header.Get(TP_SESSION_TOKEN)
-
-	td := a.sl.CheckToken(token)
-	if td == nil || td.IsServer {
+	if token := a.token(res, req); token == nil {
+		return
+	} else if token.IsServer {
 		a.sendError(res, http.StatusForbidden, statusPinResetNoServer)
 		return
 	}
@@ -82,7 +83,8 @@ func (a *Api) SendPinReset(res http.ResponseWriter, req *http.Request, vars map[
 	// first get the IMEI of the patient's handset
 	var patientConfig *portal.PatientConfig
 
-	if patientConfig, err = a.portal.GetPatientConfig(token); err != nil {
+	// actual token value is needed to call portal client: we know it does exist and is a user one
+	if patientConfig, err = a.portal.GetPatientConfig(req.Header.Get(TP_SESSION_TOKEN)); err != nil {
 		a.sendError(res, http.StatusInternalServerError, statusPinResetErr, "error getting patient config: ", err.Error())
 		return
 	}