+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 Hydrophone is the module responsible for sending emails.
 This API sends notifications to users for things like forgotten passwords, initial signup, and invitations.
 
-## Unreleased 
+## 0.10.1 - 2020-07-06
 ### Fixed 
-- PT-1406 Update the format of the message sent to patient when requesting a password reset. 
+- PT-1406 Update the format of the message sent to patient when requesting a password reset
+### Engineering
+- Review the OpenAPI documentation that was not correct on the security part
 
 ## 0.10.0 - 2020-07-02
 ### Added